requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
coreapi==2.3.3
coreschema==0.0.4
Django==4.0
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.10.0
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.0.1
psycopg2-binary==2.9.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
simplejson==3.17.6
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
zipp==3.6.0
